Updated the MPC code for calculating the bits.
In order to calculate the bits for the inner stages, party 0 sends
its share of b1 to party 1. Party 1 sends its share of b0 to party 0.
This allows party i to reconstruct bit i.
More thought is necessary to ensure that this is not a problem.

Also made the parties secret share the expansion
results in the first round of the protocol.